--- marklogic-client-api/src/main/java/com/marklogic/client/datamovement/CallManyEvent.java
@@ -1,5 +1,0 @@
-package com.marklogic.client.datamovement;
-
-public interface CallManyEvent<R> extends CallEvent, Batch<R> {
-    R[] getItems();
-}
