+++ marklogic-client-api/src/main/java/com/marklogic/client/dataservices/CallManager.java
@@ -19,10 +19,6 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.marklogic.client.DatabaseClient;
 import com.marklogic.client.SessionState;
-import com.marklogic.client.datamovement.CallBatcherBuilder;
-import com.marklogic.client.datamovement.CallEvent;
-import com.marklogic.client.datamovement.CallManyEvent;
-import com.marklogic.client.datamovement.CallOneEvent;
 import com.marklogic.client.io.marker.AbstractWriteHandle;
 import com.marklogic.client.io.marker.JSONWriteHandle;
 import org.w3c.dom.Document;
