+++ marklogic-client-api/src/main/java/com/marklogic/client/dataservices/CallManagerImpl.java
@@ -19,13 +19,8 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.marklogic.client.DatabaseClient;
 import com.marklogic.client.SessionState;
-import com.marklogic.client.datamovement.CallBatcherBuilder;
-import com.marklogic.client.datamovement.CallEvent;
-import com.marklogic.client.datamovement.CallManyEvent;
-import com.marklogic.client.datamovement.CallOneEvent;
 import com.marklogic.client.impl.BaseProxy;
 import com.marklogic.client.impl.NodeConverter;
-import com.marklogic.client.impl.RESTServices;
 import com.marklogic.client.impl.RESTServices.CallField;
 import com.marklogic.client.impl.RESTServices.MultipleAtomicCallField;
 import com.marklogic.client.impl.RESTServices.MultipleCallResponse;
