+++ marklogic-client-api/src/main/java/com/marklogic/client/dataservices/CallManyEvent.java
@@ -1,0 +1,7 @@
+package com.marklogic.client.dataservices;
+
+import com.marklogic.client.datamovement.Batch;
+
+public interface CallManyEvent<R> extends CallEvent, Batch<R> {
+    R[] getItems();
+}
