+++ marklogic-client-api/src/main/java/com/marklogic/client/dataservices/CallSuccessListener.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.client.datamovement;
+package com.marklogic.client.dataservices;
 
-public interface CallOneEvent<R> extends CallEvent {
-    R getItem();
+import com.marklogic.client.datamovement.BatchListener;
+
+public interface CallSuccessListener<E extends CallEvent> extends BatchListener<E> {
 }
